refactor lock, role, user, httpclient
